requirements.txt
backcall==0.2.0
colorama==0.4.4
cycler==0.11.0
Cython==0.29.25
debugpy==1.5.1
decorator==5.1.0
dpcpp-cpp-rt==2021.4.0
entrypoints==0.3
fonttools==4.28.2
intel-cmplr-lib-rt==2021.4.0
intel-cmplr-lic-rt==2021.4.0
intel-opencl-rt==2021.4.0
intel-openmp==2021.4.0
ipykernel==6.5.1
ipython==7.29.0
jedi==0.18.1
jupyter-client==7.1.0
jupyter-core==4.9.1
kiwisolver==1.3.2
matplotlib==3.5.0
matplotlib-inline==0.1.3
mkl==2021.4.0
mkl-fft==1.3.1
nest-asyncio==1.5.1
ninja==1.10.2.3
numpy==1.21.4
opencv-python==4.5.4.60
packaging==21.3
parso==0.8.2
pickleshare==0.7.5
Pillow==8.4.0
prompt-toolkit==3.0.22
Pygments==2.10.0
pyparsing==3.0.6
python-dateutil==2.8.2
pywin32==302
pyzmq==22.3.0
setuptools-scm==6.3.2
six==1.16.0
tbb==2021.4.0
tomli==1.2.2
torch==1.10.0+cu102
torchaudio==0.10.0+cu102
torchvision==0.11.1+cu102
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
typing_extensions==4.0.0
wcwidth==0.2.5
wget==3.2
